add test to integration test dir names
