+++ test/integration/001_simple_copy_test/models/materialized.sql
