+++ test/integration/001_simple_copy_test/test_simple_copy.py
@@ -1,29 +1,31 @@
 from test.integration.base import DBTIntegrationTest
 
-class TestVarcharWidening(DBTIntegrationTest):
+class TestSimpleCopy(DBTIntegrationTest):
 
     def setUp(self):
         DBTIntegrationTest.setUp(self)
 
-        self.run_sql_file("test/integration/002_varchar_widening/seed.sql")
+        self.run_sql_file("test/integration/001_simple_copy_test/seed.sql")
 
     @property
     def schema(self):
-        return "varchar_widening_002"
+        return "simple_copy_001"
 
     @property
     def models(self):
-        return "test/integration/002_varchar_widening/models"
+        return "test/integration/001_simple_copy_test/models"
 
-    def test_varchar_widening(self):
+    def test_simple_copy(self):
         self.run_dbt()
 
+        self.assertTablesEqual("seed","view")
         self.assertTablesEqual("seed","incremental")
         self.assertTablesEqual("seed","materialized")
 
-        self.run_sql_file("test/integration/002_varchar_widening/update.sql")
+        self.run_sql_file("test/integration/001_simple_copy_test/update.sql")
 
         self.run_dbt()
 
+        self.assertTablesEqual("seed","view")
         self.assertTablesEqual("seed","incremental")
         self.assertTablesEqual("seed","materialized")
