+++ test/integration/002_varchar_widening_test/test_varchar_widening.py
@@ -1,31 +1,29 @@
 from test.integration.base import DBTIntegrationTest
 
-class TestSimpleCopy(DBTIntegrationTest):
+class TestVarcharWidening(DBTIntegrationTest):
 
     def setUp(self):
         DBTIntegrationTest.setUp(self)
 
-        self.run_sql_file("test/integration/001_simple_copy/seed.sql")
+        self.run_sql_file("test/integration/002_varchar_widening_test/seed.sql")
 
     @property
     def schema(self):
-        return "simple_copy_001"
+        return "varchar_widening_002"
 
     @property
     def models(self):
-        return "test/integration/001_simple_copy/models"
+        return "test/integration/002_varchar_widening_test/models"
 
-    def test_simple_copy(self):
+    def test_varchar_widening(self):
         self.run_dbt()
 
-        self.assertTablesEqual("seed","view")
         self.assertTablesEqual("seed","incremental")
         self.assertTablesEqual("seed","materialized")
 
-        self.run_sql_file("test/integration/001_simple_copy/update.sql")
+        self.run_sql_file("test/integration/002_varchar_widening_test/update.sql")
 
         self.run_dbt()
 
-        self.assertTablesEqual("seed","view")
         self.assertTablesEqual("seed","incremental")
         self.assertTablesEqual("seed","materialized")
