+++ test/integration/003_simple_reference_test/models/ephemeral_copy.sql
