+++ test/integration/003_simple_reference_test/models/incremental_summary.sql
