+++ test/integration/003_simple_reference_test/models/materialized_copy.sql
