+++ test/integration/003_simple_reference_test/models/view_summary.sql
