+++ test/integration/004_simple_archive_test/update.sql
