+++ test/integration/006_simple_dependency_test/models/view_summary.sql
