+++ test/integration/006_simple_dependency_test/test_simple_dependency_with_configs.py
@@ -4,7 +4,7 @@ class BaseTestSimpleDependencyWithConfigs(DBTIntegrationTest):
 
     def setUp(self):
         DBTIntegrationTest.setUp(self)
-        self.run_sql_file("test/integration/006_simple_dependency/seed.sql")
+        self.run_sql_file("test/integration/006_simple_dependency_test/seed.sql")
 
     @property
     def schema(self):
@@ -12,7 +12,7 @@ class BaseTestSimpleDependencyWithConfigs(DBTIntegrationTest):
 
     @property
     def models(self):
-        return "test/integration/006_simple_dependency/models"
+        return "test/integration/006_simple_dependency_test/models"
 
 class TestSimpleDependencyWithConfigs(BaseTestSimpleDependencyWithConfigs):
