+++ test/integration/008_schema_tests_test/models/table_failure_copy.sql
