+++ test/integration/009_data_tests_test/models/table_copy.sql
