+++ test/integration/009_data_tests_test/test_data_tests.py
@@ -6,39 +6,45 @@ from dbt.project import read_project
 class FakeArgs(object):
     def __init__(self):
         self.threads = 1
-        self.data = False
-        self.schema = True
+        self.data = True
+        self.schema = False
         self.insert_test_results = False
 
-class TestSchemaTests(DBTIntegrationTest):
+class TestDataTests(DBTIntegrationTest):
 
     def setUp(self):
         DBTIntegrationTest.setUp(self)
-        self.run_sql_file("test/integration/008_schema_tests/seed.sql")
-        self.run_sql_file("test/integration/008_schema_tests/seed_failure.sql")
+        self.run_sql_file("test/integration/009_data_tests_test/seed.sql")
     
     @property
+    def project_config(self):
+        return {
+            "test-paths": ["test/integration/009_data_tests_test/tests"]
+        }
+
+    @property
     def schema(self):
-        return "schema_tests_008"
+        return "data_tests_009"
 
     @property
     def models(self):
-        return "test/integration/008_schema_tests/models"
+        return "test/integration/009_data_tests_test/models"
 
-    def run_schema_validations(self):
+    def run_data_validations(self):
         project = read_project('dbt_project.yml')
         args = FakeArgs()
 
         test_task = TestTask(args, project)
         return test_task.run()
 
-    def test_schema_tests(self):
+    def test_data_tests(self):
         self.run_dbt()
-        test_results = self.run_schema_validations()
+        test_results = self.run_data_validations()
+
 
         for result in test_results:
             # assert that all deliberately failing tests actually fail
-            if 'failure' in result.model.name:
+            if 'fail' in result.model.name:
                 self.assertFalse(result.errored)
                 self.assertFalse(result.skipped)
                 self.assertTrue(result.status > 0)
