+++ test/integration/009_data_tests_test/tests/fail_email_is_always_null.sql
